Tweak so we don't rely on brew being in the path
Most of the issues people are having stem from not reloading the PATH
env variable once /opt/homebrew/bin is added.

To mitigate that, we check for the existence of `/opt/homebrew/bin/brew`
instead of the output of `which brew`.

Additionally, other installation commands use the full path to brew.